+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tryabby/angular
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/core@5.2.0
+
 ## 2.0.6
 
 ### Patch Changes