+++ packages/next/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tryabby/next
 
+## 5.1.0
+
+### Minor Changes
+
+- add helper to retrieve all flags/config variables
+
+### Patch Changes
+
+- Updated dependencies
+  - @tryabby/react@5.1.0
+
 ## 5.0.6
 
 ### Patch Changes